requirements.txt
langchain>=0.0.200
openai>=1.0.0
faiss-cpu>=1.7.4
langchain-community>=0.0.10
tiktoken>=0.5.0
langchain-openai>=0.0.1
pandas>=2.0.0
tqdm>=4.66.0
tenacity>=8.2.0
nltk>=3.8.0
scikit-learn>=1.3.0
python-dotenv>=1.0.0
